# openapi.yaml
openapi: 3.1.0
info:
  title: Bakame API
  version: 1.0.0
  description: |
    Bakame - Production-grade AI platform API

    All endpoints except /health require Bearer token authentication.

    Response Format:
    - Success: { data: <payload>, meta: { requestId: string } }
    - Error: { error: { code: string, message: string, requestId: string } }
servers:
  - url: /api/v1
    description: API v1
...
paths:
...
  /memories:
    get:
      tags: [Memories]
      summary: List or search memories
      operationId: listMemories
      security:
        - bearerAuth: []
      parameters:
        - name: search
          in: query
          description: Search query (uses semantic search)
          schema:
            type: string
        - name: category
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: cursor
          in: query
          schema:
            type: string
      responses:
        '200':
          description: List of memories

    post:
      tags: [Memories]
      summary: Create memory
      operationId: createMemory
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [content]
              properties:
                content:
                  type: string
                category:
                  type: string
                importance:
                  type: integer
                  minimum: 1
                  maximum: 10
      responses:
        '201':
          description: Memory created
        '400':
          $ref: '#/components/responses/ValidationError'
  /memories/{id}:
    get:
      tags: [Memories]
      summary: Get memory
      operationId: getMemory
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/MemoryId'
      responses:
        '200':
          description: Memory details
        '404':
          $ref: '#/components/responses/NotFound'

    patch:
      tags: [Memories]
      summary: Update memory
      operationId: updateMemory
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/MemoryId'
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                content:
                  type: string
                importance:
                  type: integer
                  minimum: 1
                  maximum: 10
      responses:
        '200':
          description: Memory updated
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags: [Memories]
      summary: Archive memory
      operationId: archiveMemory
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/MemoryId'
      responses:
        '204':
          description: Memory archived
...
  /tools:
    get:
      tags: [Tools]
      summary: List available tools
      operationId: listTools
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Available tools
  /tools/{id}:
    get:
      tags: [Tools]
      summary: Get tool details
      operationId: getTool
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Tool details
        '404':
          $ref: '#/components/responses/NotFound'
  /subscription:
    get:
      tags: [Subscription]
      summary: Get subscription
      operationId: getSubscription
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Subscription details
  /subscription/usage:
    get:
      tags: [Subscription]
      summary: Get usage summary
      operationId: getUsage
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Usage summary
  /subscription/entitlements:
    get:
      tags: [Subscription]
      summary: Get entitlements
      operationId: getEntitlements
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Feature entitlements
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    MemoryId:
      name: id
      in: path
      required: true
      schema:
        type: string
...
  responses:
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    ValidationError:
      description: Validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, requestId]
          properties:
            code:
              type: string
              enum:
                - VALIDATION_ERROR
                - UNAUTHORIZED
                - PERMISSION_DENIED
                - NOT_FOUND
                - RATE_LIMITED
                - INTERNAL_ERROR
            message:
              type: string
            requestId:
              type: string
...